src/master/requirements.txt
anyio==3.7.1
bidict==0.22.1
click==8.1.5
h11==0.14.0
httptools==0.6.0
idna==3.4
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
PyYAML==6.0
sniffio==1.3.0
uvicorn==0.23.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
psycopg2==2.9.7
